fix build
git-svn-id: http://svn.openzap.org/svn/openzap/trunk@647 a93c3328-9c30-0410-af19-c9cd2b2d52af

+++ libs/freetdm/src/testcid.c
@@ -67,7 +67,7 @@ int main(int argc, char *argv[])
 		zap_fsk_modulator_init(&fsk_trans, FSK_BELL202, 8000, &fsk_data, -14, 180, 5, 300, my_write_sample, &foo);
 		zap_fsk_modulator_send_all((&fsk_trans));
 
-		printf("%lu %d %d\n", fsk_data.dlen, foo.wrote, fsk_trans.est_bytes);
+		printf("%u %d %d\n", (unsigned) fsk_data.dlen, foo.wrote, fsk_trans.est_bytes);
 
 		if (fd > -1) {
 			close (fd);
